fix contest worker lifecycle test
